requirementes.txt
requests
pandas
google-cloud-bigquery
functions-framework==3.*
db-dtypes